IDGEN-67 Follow Up (#51)
