+++ api/src/test/resources/org/openmrs/module/idgen/include/TestData.xml
@@ -31,5 +31,5 @@
    <IDGEN_LOG_ENTRY ID="2" SOURCE="1" GENERATED_BY="1" IDENTIFIER="100HH9" COMMENT="New Visit" DATE_GENERATED="2017-10-03 12:00:00.489" />
    <IDGEN_LOG_ENTRY ID="3" SOURCE="1" GENERATED_BY="2" IDENTIFIER="100H9T" DATE_GENERATED="2017-10-01 12:00:00.489" />
    <IDGEN_LOG_ENTRY ID="4" SOURCE="2" GENERATED_BY="3" IDENTIFIER="100H28" COMMENT="Demo Data" DATE_GENERATED="2017-10-02 12:00:00.489" />
-   <IDGEN_LOG_ENTRY ID="5" SOURCE="3" GENERATED_BY="4" IDENTIFIER="100GYF" DATE_GENERATED="2017-09-20 12:00:00.489" />
+   <IDGEN_LOG_ENTRY ID="5" SOURCE="3" GENERATED_BY="4" IDENTIFIER="100GYF" DATE_GENERATED="2017-09-20 12:00:00.489" />  
 </dataset>
