+++ omod/src/test/java/org/openmrs/module/idgen/rest/resource/LogEntryResourceTest.java
@@ -6,7 +6,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.idgen.web.resource;
+package org.openmrs.module.idgen.rest.resource;
 
 import org.junit.Before;
 import org.openmrs.api.context.Context;
